PHP 8.4 compatiblity: use hash() for hashing
PHP 8.4 will deprecate md5(), md5_file(), sha1(), and sha1_file(). The
recommendation is to use hash() and hash_file() which is safe because
ext/hash is mandatory since PHP 7.4.0.

For now, we are going to be using a shim. Future versions will move to
using hash() and hash_file() directly.